Creo el paquete AlgoCraft para resolver dependencias
