+++ src/AlgoCraft/Desgaste.java
@@ -1,4 +1,4 @@
-package Objetos;
+package AlgoCraft;
 
 public  abstract class Desgaste {
     public abstract int desgastarDada(int durabilidad, int fuerza);
