+++ src/AlgoCraft/Diamante.java
@@ -1,4 +1,4 @@
-package Objetos;
+package AlgoCraft;
 
 public class Diamante extends Material {
     public Diamante(){
