+++ src/AlgoCraft/Hacha.java
@@ -1,28 +1,26 @@
-package Objetos;
+package AlgoCraft;
 
-public class Pico extends Herramienta {
+public class Hacha extends Herramienta {
 
-    public Pico(MaterialDeConstruccion unMaterial){
+    public Hacha(MaterialDeConstruccion unMaterial){
         this.material = unMaterial;
         this.durabilidad = unMaterial.durabilidadDeConstruccion(this);
         this.fuerza = unMaterial.fuerzaDeConstruccion(this);
         this.desgaste = unMaterial.desgasteDeConstruccion(this);
     }
 
-    @Override
     public void golpear(Material material) {
-
         durabilidad = desgaste.desgastarDada(durabilidad, fuerza);
     }
 
     @Override
     public boolean esUnHacha() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean esUnPico() {
-        return true;
+        return false;
     }
 
     @Override
