+++ src/AlgoCraft/MaterialDeConstruccion.java
@@ -1,4 +1,4 @@
-package Objetos;
+package AlgoCraft;
 
 public  abstract class MaterialDeConstruccion extends Material {
