+++ test/CasilleroTest.java
@@ -1,3 +1,5 @@
+import AlgoCraft.Casillero;
+
 public class CasilleroTest {
     private Casillero casillero = new Casillero();
