+++ test/MapaTest.java
@@ -1,3 +1,4 @@
+import AlgoCraft.Mapa;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
